changes to project setup new package util first version of beat position detection

+++ at.cp.jku.teaching.amprocessing/src/at/cp/jku/teaching/amprocessing/project/BeatDetector.java
@@ -4,6 +4,6 @@ import java.util.List;
 
 public interface BeatDetector {
 
-	List<Double> analyze(double[] onsetFunction);
+	List<Integer> analyze(double[] onsetFunction);
 
 }
